requirements.txt
absl-py==1.1.0
astor==0.8.1
atari-py==0.2.9
certifi==2022.6.15
click==8.1.3
cloudpickle==1.3.0
cycler==0.11.0
dill==0.3.5.1
docopt==0.6.2
future==0.18.2
gast==0.2.2
gitdb==4.0.9
GitPython==3.1.27
glob2==0.7
google-pasta==0.2.0
grpcio==1.47.0
gym==0.17.2
gym-notices==0.0.7
h5py==3.7.0
importlib-metadata==4.12.0
joblib==1.1.0
jsonpickle==0.9.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
Markdown==3.3.7
matplotlib==3.0.3
memory-profiler==0.60.0
mock==4.0.3
mpi4py==3.0.3
munch==2.5.1.dev12
numpy==1.18.5
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
pandas==1.0.5
Pillow==9.2.0
progressbar2==4.0.0
protobuf==3.20.0
psutil==5.9.1
py-cpuinfo==8.0.0
pygame==1.9.5
pyglet==1.5.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.3.3
pytz==2022.1
pyzmq==23.2.0
sacred==0.7.4
scipy==1.5.0
seaborn==0.9.0
setuptools==61.2.0
six==1.16.0
smmap==5.0.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
torch==1.12.0
tqdm==4.64.0
typing_extensions==4.3.0
Werkzeug==2.1.2
wheel==0.37.1
wrapt==1.14.1
zipp==3.8.0
zmq==0.0.0
networkx
scikit-learn
